Tuplet is a private class used only by the BlockingQueueTests.
