+++ src/tests/DreamMisc/test.mindtouch.dream.csproj
@@ -112,7 +112,6 @@
     <Compile Include="Collections\WorkStealingDequeTests.cs" />
     <Compile Include="CookieTests.cs" />
     <Compile Include="MimeTypeTests.cs" />
-    <Compile Include="Tuplet.cs" />
     <Compile Include="XUriParser-Test.cs" />
     <Compile Include="DataUpdaterTests.cs" />
     <Compile Include="DreamFeatureTests.cs" />
